Added reset etc
In this commit I added reset to project 5,
I also changed so we only run project 5 from start,
i.e. we skip the menu, this way we can run the simulator
using -batchmode flag, this runs the simulator
in an headless mode and thus faster.
I also added a possibility to speed up the simulator,
using unitys timeScale.